PARQUET-400: Fix review nits.
