+++ parquet-hadoop/src/test/java/org/apache/parquet/hadoop/util/TestHadoop1ByteBufferReads.java
@@ -24,13 +24,12 @@ import org.apache.parquet.hadoop.TestUtils;
 import org.junit.Assert;
 import org.junit.Test;
 import java.io.EOFException;
-import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.concurrent.Callable;
 
 import static org.apache.parquet.hadoop.util.MockInputStream.TEST_ARRAY;
 
-public class TestByteBufferReads {
+public class TestHadoop1ByteBufferReads {
 
   private static final ThreadLocal<byte[]> TEMP = new ThreadLocal<byte[]>() {
     @Override
@@ -39,31 +38,8 @@ public class TestByteBufferReads {
     }
   };
 
-  /**
-   * This mimics ByteBuffer reads from streams in Hadoop 2
-   */
-  private static class MockBufferReader implements H2SeekableInputStream.Reader {
-    private final FSDataInputStream stream;
-
-    public MockBufferReader(FSDataInputStream stream) {
-      this.stream = stream;
-    }
-
-    @Override
-    public int read(ByteBuffer buf) throws IOException {
-      // this is inefficient, but simple for correctness tests of
-      // readFully(ByteBuffer)
-      byte[] temp = new byte[buf.remaining()];
-      int bytesRead = stream.read(temp, 0, temp.length);
-      if (bytesRead > 0) {
-        buf.put(temp, 0, bytesRead);
-      }
-      return bytesRead;
-    }
-  }
-
   @Test
-  public void testH1HeapRead() throws Exception {
+  public void testHeapRead() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocate(20);
 
     FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -82,7 +58,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapSmallBuffer() throws Exception {
+  public void testHeapSmallBuffer() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocate(5);
 
     FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -101,7 +77,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapSmallReads() throws Exception {
+  public void testHeapSmallReads() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocate(10);
 
     FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
@@ -132,7 +108,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapPosition() throws Exception {
+  public void testHeapPosition() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocate(20);
     readBuffer.position(10);
     readBuffer.mark();
@@ -158,7 +134,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapLimit() throws Exception {
+  public void testHeapLimit() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocate(20);
     readBuffer.limit(8);
 
@@ -183,7 +159,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapPositionAndLimit() throws Exception {
+  public void testHeapPositionAndLimit() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocate(20);
     readBuffer.position(5);
     readBuffer.limit(13);
@@ -210,7 +186,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectRead() throws Exception {
+  public void testDirectRead() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocateDirect(20);
 
     FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -229,7 +205,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectSmallBuffer() throws Exception {
+  public void testDirectSmallBuffer() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocateDirect(5);
 
     FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -248,7 +224,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectSmallReads() throws Exception {
+  public void testDirectSmallReads() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
 
     FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
@@ -279,7 +255,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectPosition() throws Exception {
+  public void testDirectPosition() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocateDirect(20);
     readBuffer.position(10);
     readBuffer.mark();
@@ -305,7 +281,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectLimit() throws Exception {
+  public void testDirectLimit() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocate(20);
     readBuffer.limit(8);
 
@@ -330,7 +306,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectPositionAndLimit() throws Exception {
+  public void testDirectPositionAndLimit() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocateDirect(20);
     readBuffer.position(5);
     readBuffer.limit(13);
@@ -357,7 +333,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectSmallTempBufferSmallReads() throws Exception {
+  public void testDirectSmallTempBufferSmallReads() throws Exception {
     byte[] temp = new byte[2]; // this will cause readDirectBuffer to loop
 
     ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
@@ -393,7 +369,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectSmallTempBufferWithPositionAndLimit() throws Exception {
+  public void testDirectSmallTempBufferWithPositionAndLimit() throws Exception {
     byte[] temp = new byte[2]; // this will cause readDirectBuffer to loop
 
     ByteBuffer readBuffer = ByteBuffer.allocateDirect(20);
@@ -422,7 +398,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapReadFullySmallBuffer() throws Exception {
+  public void testHeapReadFullySmallBuffer() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocate(8);
 
     FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -441,7 +417,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapReadFullyLargeBuffer() throws Exception {
+  public void testHeapReadFullyLargeBuffer() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocate(20);
 
     final FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -460,7 +436,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapReadFullyJustRight() throws Exception {
+  public void testHeapReadFullyJustRight() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocate(10);
 
     final FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -481,7 +457,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapReadFullySmallReads() throws Exception {
+  public void testHeapReadFullySmallReads() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocate(10);
 
     final FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
@@ -500,7 +476,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapReadFullyPosition() throws Exception {
+  public void testHeapReadFullyPosition() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocate(10);
     readBuffer.position(3);
     readBuffer.mark();
@@ -521,7 +497,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapReadFullyLimit() throws Exception {
+  public void testHeapReadFullyLimit() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocate(10);
     readBuffer.limit(7);
 
@@ -551,7 +527,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1HeapReadFullyPositionAndLimit() throws Exception {
+  public void testHeapReadFullyPositionAndLimit() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocate(10);
     readBuffer.position(3);
     readBuffer.limit(7);
@@ -583,7 +559,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectReadFullySmallBuffer() throws Exception {
+  public void testDirectReadFullySmallBuffer() throws Exception {
     ByteBuffer readBuffer = ByteBuffer.allocateDirect(8);
 
     FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -602,7 +578,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectReadFullyLargeBuffer() throws Exception {
+  public void testDirectReadFullyLargeBuffer() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocateDirect(20);
 
     final FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -627,7 +603,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectReadFullyJustRight() throws Exception {
+  public void testDirectReadFullyJustRight() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
 
     final FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
@@ -648,7 +624,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectReadFullySmallReads() throws Exception {
+  public void testDirectReadFullySmallReads() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
 
     final FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
@@ -667,7 +643,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectReadFullyPosition() throws Exception {
+  public void testDirectReadFullyPosition() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
     readBuffer.position(3);
     readBuffer.mark();
@@ -688,7 +664,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectReadFullyLimit() throws Exception {
+  public void testDirectReadFullyLimit() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
     readBuffer.limit(7);
 
@@ -718,7 +694,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectReadFullyPositionAndLimit() throws Exception {
+  public void testDirectReadFullyPositionAndLimit() throws Exception {
     final ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
     readBuffer.position(3);
     readBuffer.limit(7);
@@ -750,7 +726,7 @@ public class TestByteBufferReads {
   }
 
   @Test
-  public void testH1DirectReadFullySmallTempBufferWithPositionAndLimit() throws Exception {
+  public void testDirectReadFullySmallTempBufferWithPositionAndLimit() throws Exception {
     byte[] temp = new byte[2]; // this will cause readFully to loop
 
     final ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
@@ -775,354 +751,6 @@ public class TestByteBufferReads {
     readBuffer.position(7);
     readBuffer.limit(10);
     H1SeekableInputStream.readFullyDirectBuffer(hadoopStream, readBuffer, temp);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.reset();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 7), readBuffer);
-  }
-
-  @Test
-  public void testH2HeapReadFullySmallBuffer() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocate(8);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(8, readBuffer.position());
-    Assert.assertEquals(8, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(8, readBuffer.position());
-    Assert.assertEquals(8, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 8), readBuffer);
-  }
-
-  @Test
-  public void testH2HeapReadFullyLargeBuffer() throws Exception {
-    final ByteBuffer readBuffer = ByteBuffer.allocate(20);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
-    final MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    TestUtils.assertThrows("Should throw EOFException",
-        EOFException.class, new Callable() {
-          @Override
-          public Object call() throws Exception {
-            H2SeekableInputStream.readFully(reader, readBuffer);
-            return null;
-          }
-        });
-
-    // NOTE: This behavior differs from readFullyHeapBuffer because direct uses
-    // several read operations that will read up to the end of the input. This
-    // is a correct value because the bytes in the buffer are valid. This
-    // behavior can't be implemented for the heap buffer without using the read
-    // method instead of the readFully method on the underlying
-    // FSDataInputStream.
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(20, readBuffer.limit());
-  }
-
-  @Test
-  public void testH2HeapReadFullyJustRight() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocate(10);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    // reads all of the bytes available without EOFException
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    // trying to read 0 more bytes doesn't result in EOFException
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY), readBuffer);
-  }
-
-  @Test
-  public void testH2HeapReadFullySmallReads() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocate(10);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY), readBuffer);
-  }
-
-  @Test
-  public void testH2HeapReadFullyPosition() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocate(10);
-    readBuffer.position(3);
-    readBuffer.mark();
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.reset();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 7), readBuffer);
-  }
-
-  @Test
-  public void testH2HeapReadFullyLimit() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocate(10);
-    readBuffer.limit(7);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(7, readBuffer.position());
-    Assert.assertEquals(7, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(7, readBuffer.position());
-    Assert.assertEquals(7, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 7), readBuffer);
-
-    readBuffer.position(7);
-    readBuffer.limit(10);
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY), readBuffer);
-  }
-
-  @Test
-  public void testH2HeapReadFullyPositionAndLimit() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocate(10);
-    readBuffer.position(3);
-    readBuffer.limit(7);
-    readBuffer.mark();
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(7, readBuffer.position());
-    Assert.assertEquals(7, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(7, readBuffer.position());
-    Assert.assertEquals(7, readBuffer.limit());
-
-    readBuffer.reset();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 4), readBuffer);
-
-    readBuffer.position(7);
-    readBuffer.limit(10);
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.reset();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 7), readBuffer);
-  }
-
-  @Test
-  public void testH2DirectReadFullySmallBuffer() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocateDirect(8);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(8, readBuffer.position());
-    Assert.assertEquals(8, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(8, readBuffer.position());
-    Assert.assertEquals(8, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 8), readBuffer);
-  }
-
-  @Test
-  public void testH2DirectReadFullyLargeBuffer() throws Exception {
-    final ByteBuffer readBuffer = ByteBuffer.allocateDirect(20);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
-    final MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    TestUtils.assertThrows("Should throw EOFException",
-        EOFException.class, new Callable() {
-          @Override
-          public Object call() throws Exception {
-            H2SeekableInputStream.readFully(reader, readBuffer);
-            return null;
-          }
-        });
-
-    // NOTE: This behavior differs from readFullyHeapBuffer because direct uses
-    // several read operations that will read up to the end of the input. This
-    // is a correct value because the bytes in the buffer are valid. This
-    // behavior can't be implemented for the heap buffer without using the read
-    // method instead of the readFully method on the underlying
-    // FSDataInputStream.
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(20, readBuffer.limit());
-  }
-
-  @Test
-  public void testH2DirectReadFullyJustRight() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream());
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    // reads all of the bytes available without EOFException
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    // trying to read 0 more bytes doesn't result in EOFException
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY), readBuffer);
-  }
-
-  @Test
-  public void testH2DirectReadFullySmallReads() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY), readBuffer);
-  }
-
-  @Test
-  public void testH2DirectReadFullyPosition() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
-    readBuffer.position(3);
-    readBuffer.mark();
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.reset();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 7), readBuffer);
-  }
-
-  @Test
-  public void testH2DirectReadFullyLimit() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
-    readBuffer.limit(7);
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
-    H2SeekableInputStream.Reader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(7, readBuffer.position());
-    Assert.assertEquals(7, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(7, readBuffer.position());
-    Assert.assertEquals(7, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 7), readBuffer);
-
-    readBuffer.position(7);
-    readBuffer.limit(10);
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(10, readBuffer.position());
-    Assert.assertEquals(10, readBuffer.limit());
-
-    readBuffer.flip();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY), readBuffer);
-  }
-
-  @Test
-  public void testH2DirectReadFullyPositionAndLimit() throws Exception {
-    ByteBuffer readBuffer = ByteBuffer.allocateDirect(10);
-    readBuffer.position(3);
-    readBuffer.limit(7);
-    readBuffer.mark();
-
-    FSDataInputStream hadoopStream = new FSDataInputStream(new MockInputStream(2, 3, 3));
-    MockBufferReader reader = new MockBufferReader(hadoopStream);
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(7, readBuffer.position());
-    Assert.assertEquals(7, readBuffer.limit());
-
-    H2SeekableInputStream.readFully(reader, readBuffer);
-    Assert.assertEquals(7, readBuffer.position());
-    Assert.assertEquals(7, readBuffer.limit());
-
-    readBuffer.reset();
-    Assert.assertEquals("Buffer contents should match",
-        ByteBuffer.wrap(TEST_ARRAY, 0, 4), readBuffer);
-
-    readBuffer.position(7);
-    readBuffer.limit(10);
-    H2SeekableInputStream.readFully(reader, readBuffer);
     Assert.assertEquals(10, readBuffer.position());
     Assert.assertEquals(10, readBuffer.limit());
